+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Breaking change
 - Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Installation#how-to-delete-and-re-add-the-solarcharger-integration).
 ### Fix
+- Work-around error with not able to reconfigure OCPP settings. 
 ### New feature
 - Added [power source](https://github.com/flashg1/SolarCharger/wiki/User-guide#source-limit-output-power) support.
 ### Improvement